Refactor: 랜덤값 생성, Integer값 변환을 Util 클래스로 분리
